actor-critic/requirements.txt
torch>=2.0.0
pandas>=1.5.0
pandas-ta>=0.3.0
numpy>=1.20.0
gym>=0.26.0
yfinance>=0.2.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0
pyyaml>=6.0.0
scikit-learn>=1.0.0
optuna>=3.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0 